refactoring ipe_image support routines
